rename figure.py to figure_manager.py
