+++ examples/01_basic_functionality.py
@@ -3,7 +3,7 @@ Example 1: Basic Functionality - Core plot types without visual encoding.
 Demonstrates fundamental plotting capabilities across all basic plot types.
 """
 
-from dr_plotter.figure import FigureManager
+from dr_plotter.figure_manager import FigureManager
 from dr_plotter.plot_config import PlotConfig
 from dr_plotter.scripting.utils import setup_arg_parser, show_or_save_plot
 from dr_plotter.scripting.verif_decorators import verify_plot, inspect_plot_properties
