+++ examples/04_specialized_plots.py
@@ -3,7 +3,7 @@ Example 4: Specialized Plots - Heatmap, contour, and distribution plot types.
 Demonstrates dr_plotter's handling of specialized plot types with unique data requirements.
 """
 
-from dr_plotter.figure import FigureManager
+from dr_plotter.figure_manager import FigureManager
 from dr_plotter.plot_config import PlotConfig
 from dr_plotter.scripting.utils import setup_arg_parser, show_or_save_plot
 from dr_plotter.scripting.verif_decorators import verify_plot, inspect_plot_properties
