+++ examples/19_ml_dashboard.py
@@ -3,7 +3,7 @@ Example 19: ML Dashboard Dashboard - Real-world use case.
 Complete ML experiment visualization with multiple metrics and hyperparameters.
 """
 
-from dr_plotter.figure import FigureManager
+from dr_plotter.figure_manager import FigureManager
 from dr_plotter.plot_config import PlotConfig
 from dr_plotter.scripting.utils import setup_arg_parser, show_or_save_plot
 from dr_plotter.scripting.verif_decorators import verify_plot, inspect_plot_properties
