+++ examples/22_faceted_training_curves_refactored.py
@@ -4,7 +4,7 @@ import sys
 import time
 import pandas as pd
 import matplotlib.pyplot as plt
-from dr_plotter.figure import FigureManager
+from dr_plotter.figure_manager import FigureManager
 from dr_plotter.plot_config import PlotConfig
 from dr_plotter.scripting.datadec_utils import get_datadec_functions
