+++ examples/27_multi_metric_plotting.py
@@ -3,7 +3,7 @@ Example 6: Multi-Metric Plotting - METRICS constant.
 Demonstrates plotting multiple y-columns with the METRICS constant.
 """
 
-from dr_plotter.figure import FigureManager
+from dr_plotter.figure_manager import FigureManager
 from dr_plotter.plot_config import PlotConfig
 from dr_plotter.scripting.utils import setup_arg_parser, show_or_save_plot
 from dr_plotter.scripting.verif_decorators import verify_plot, inspect_plot_properties
