+++ src/dr_plotter/__init__.py
@@ -1,5 +1,5 @@
 from . import consts
 from .faceting_config import FacetingConfig
-from .figure import FigureManager
+from .figure_manager import FigureManager
 
 __version__ = "0.1.0"
